added argument "no_hello_video" to prevent the build and install of h… (#36)
* added argument "no_hello_video" to prevent the build and install of hello_video

+++ install.sh
@@ -9,19 +9,27 @@ if [ "$(id -u)" != "0" ]; then
   exit 1
 fi
 
+
 echo "Installing dependencies..."
 echo "=========================="
-apt update && apt -y install git build-essential python3-dev python3 python3-pip python3-pygame supervisor omxplayer
+apt update && apt -y install python3 python3-pip python3-pygame supervisor omxplayer ntfs-3g exfat-fuse
 
-echo "Installing hello_video..."
-echo "========================="
-git clone https://github.com/adafruit/pi_hello_video.git
-cd pi_hello_video
-./rebuild.sh
-cd hello_video
-make install
-cd ../..
-rm -rf pi_hello_video
+if [ "$*" != "no_hello_video" ]
+then
+	echo "Installing hello_video..."
+	echo "========================="
+	apt -y install git build-essential python3-dev
+	git clone https://github.com/adafruit/pi_hello_video.gi
+	cd pi_hello_video
+	./rebuild.sh
+	cd hello_video
+	make install
+	cd ../..
+	rm -rf pi_hello_video
+else
+    echo "hello_video was not installed"
+    echo "=========================="
+fi
 
 echo "Installing video_looper program..."
 echo "=================================="
